requirements.txt
streamlit==1.24.0
requests==2.31.0
python-dotenv==1.0.1
pypdf==4.0.1
openai
langgraph 
langsmith
langchain
langchain[openai]
langchain-tavily
langgraph-supervisor
langchain-google-community[calendar]